General API reference

Scheduler API reference

ParameterSchedulers.Scheduler — Type

Scheduler{T, O, F}(schedule::AbstractSchedule, opt, update_func)
Scheduler(schedule, opt; update_func = (o, s) -> (o.eta = s))

Wrap a schedule and opt together with a Scheduler. The schedule is iterated on every call to Flux.apply!. The Scheduler can be used anywhere a Flux optimizer is used.

By default, the learning rate (i.e. opt.eta) is scheduled. Set update_func = (opt, schedule_val) -> ... to schedule an alternate field. If opt does not have a field eta, then there is no default behavior (you must manually set update_func).

Arguments

schedule: the schedule to use

opt: a Flux optimizer

update_func: a mutating function of with inputs (optim, param) that mutates optim's fields based on the current param value

Examples

# cosine annealing schedule for Descent
julia> s = CosAnneal(λ0 = 0.1, λ1 = 0.8, period = 10);

julia> opt = Scheduler(s, Descent())
Scheduler(CosAnneal{Float64,Int64}(0.1, 0.8, 10), Descent(0.1))

# schedule the momentum term of Momentum
julia> opt = Scheduler(s, Momentum(); update_func = (o, s) -> o.rho = s)
Scheduler(CosAnneal{Float64,Int64}(0.1, 0.8, 10), Momentum(0.01, 0.9, IdDict{Any,Any}()))

State schedule API reference

ParameterSchedulers.Stateful — Type

Stateful{T, S}
Stateful(schedule::T; advance = state -> true)

Create a stateful iterator around schedule. Pass in a predicate, advance(state), to conditionally control iteration. See also ParameterSchedulers.next! and ParameterSchedulers.reset!.

ParameterSchedulers.next! — Function

next!(iter::Stateful)

Advance iter by one iteration (if iter.advance(state) == true) and return the next value. See also ParameterSchedulers.Stateful.

ParameterSchedulers.reset! — Function

reset!(iter::Stateful)

Reset iter to its initial state. See also ParameterSchedulers.Stateful.

Utility functions API reference

ParameterSchedulers.reverse — Method

reverse(f, period)

Return a reverse function such that reverse(f, period)(t) == f(period - t).

ParameterSchedulers.symmetric — Method

symmetric(f, period)

Return a symmetric function such that for t ∈ [1, period / 2), the symmetric function evaluates to f(t), and when t ∈ [period / 2, period), the symmetric functions evaluates to f(period - t).